chore: move register into registers in storage
