+++ safenode/src/protocol/storage/registers/register/metadata.rs
